+++ service/controller/managementcluster/resource.go
@@ -159,6 +159,7 @@ func newResources(config resourcesConfig) ([]resource.Interface, error) {
 	{
 		c := ciliumnetpol.Config{
 			DynamicK8sClient: config.DynamicK8sClient,
+			Proxy:            config.Proxy,
 			Logger:           config.Logger,
 		}